+++ ext-src/NexusExplorer.ts
@@ -15,306 +15,14 @@
  */
 import * as vscode from "vscode";
 import * as path from "path";
-import * as fs from "fs";
 import * as _ from "lodash";
-import * as request from "request";
 import * as dependencyTree from "dependency-tree";
 
 import {
   ComponentInfoPanel,
-  ComponentEntry,
-  PolicyViolation
+  ComponentEntry
 } from "./ComponentInfoPanel";
-import { PackageDependencies } from "./packages/PackageDependencies";
-import { MavenDependencies } from "./packages/maven/MavenDependencies";
-import { NpmDependencies } from "./packages/npm/NpmDependencies";
-import { MavenCoordinate } from "./packages/maven/MavenCoordinate";
-import { NpmCoordinate } from "./packages/npm/NpmCoordinate";
-
-enum DependencyType {
-  NPM = "NPM",
-  Maven = "Maven",
-}
-
-export class IqComponentModel {
-  components: Array<ComponentEntry> = [];
-  coordsToComponent: Map<string, ComponentEntry> = new Map<
-    string,
-    ComponentEntry
-  >();
-
-  // TODO make these configurable???
-  //
-  readonly evaluationPollDelayMs = 2000;
-
-  constructor(
-    readonly url: string,
-    private user: string,
-    private password: string,
-    private applicationPublicId: string,
-    private getmaximumEvaluationPollAttempts: number
-  ) {}
-
-  public getContent(resource: vscode.Uri): Thenable<string> {
-    // TODO get the HTML doc for webview
-    return new Promise((c, e) => "my stubbed content entry");
-  }
-
-  public async evaluateComponents() {
-    console.debug("evaluateComponents");
-    await this.performIqScan();
-  }
-
-  private determineWorkspaceDependencyType(workspaceRoot: string): DependencyType {
-    const packageJsonPath = path.join(workspaceRoot, "package.json");
-    const pomFilePath = path.join(workspaceRoot, "pom.xml");
-
-    const packageJsonExists: boolean = this.pathExists(packageJsonPath);
-    const pomFileExists: boolean =  this.pathExists(pomFilePath); 
-
-    if (packageJsonExists){
-      return DependencyType.NPM;
-    }
-    else if (pomFileExists) {
-      return DependencyType.Maven;
-    }
-    else{
-      throw new TypeError("Workspace has no package.json or pom.xml");
-    }
-  }
-
-  private async packageForIQ(workspaceRoot: string, type: DependencyType): Promise<PackageDependencies> {
-    switch (type) {
-      case DependencyType.Maven:
-        let mavenDependencies = new MavenDependencies();
-        mavenDependencies.packageForIq(workspaceRoot);
-        return mavenDependencies;
-      case DependencyType.NPM:
-        let npmDependencies = new NpmDependencies();
-        npmDependencies.packageForIq(workspaceRoot);
-        return npmDependencies;
-      default:
-        throw new TypeError("Functionality not implemented");
-    }
-  }
-
-  private async packageAndConvertToNexusFormat(dependencyType: DependencyType, workspaceRoot: string) {
-    let data = undefined;
-    let items = await this.packageForIQ(workspaceRoot, dependencyType);
-    data = items.convertToNexusFormat();
-    this.components = items.toComponentEntries(data);
-
-    return data;
-  }
-
-  private async performIqScan() {
-    try {
-      const workspaceRoot = vscode.workspace.rootPath;
-      if (workspaceRoot === undefined) {
-        throw new TypeError("No workspace opened");
-      }
-
-      const dependencyType = this.determineWorkspaceDependencyType(workspaceRoot);
-
-      console.debug("Project dependency type:" + dependencyType)
-
-      let data = this.packageAndConvertToNexusFormat(dependencyType, workspaceRoot);
-
-      if (undefined == data) {
-        throw new RangeError("Attempted to generated dependency list but received an empty collection. NexusIQ will not be invoked for this project.");
-      }
-
-      console.debug("getting applicationInternalId", this.applicationPublicId);
-      let response = await this.getApplicationId(this.applicationPublicId);
-
-      let appRep = JSON.parse(response as string);
-      console.debug("appRep", appRep);
-
-      let applicationInternalId = appRep.applications[0].id;
-      console.debug("applicationInternalId", applicationInternalId);
-
-      let resultId = await this.submitToIqForEvaluation(
-        data,
-        applicationInternalId as string
-      );
-
-      console.debug("report", resultId);
-      let resultDataString = await this.asyncPollForEvaluationResults(
-        applicationInternalId as string,
-        resultId as string
-      );
-      let resultData = JSON.parse(resultDataString as string);
-
-      console.debug(`Received results from IQ scan:`, resultData);
-      for (let resultEntry of resultData.results) {
-
-        let componentEntry: ComponentEntry | undefined = undefined;
-        if (dependencyType === DependencyType.NPM) {
-          let coordinates = resultEntry.component.componentIdentifier
-            .coordinates as NpmCoordinate;
-          componentEntry = this.coordsToComponent.get(
-            coordinates.asCoordinates()
-          );
-        }
-        else if (dependencyType === DependencyType.Maven){
-          let coordinates = resultEntry.component.componentIdentifier
-            .coordinates as MavenCoordinate;
-          componentEntry = this.coordsToComponent.get(
-            coordinates.asCoordinates()
-          );
-        }
-        componentEntry!.policyViolations = resultEntry.policyData
-          .policyViolations as Array<PolicyViolation>;
-        componentEntry!.hash = resultEntry.component.hash;
-        componentEntry!.nexusIQData = resultEntry;
-      }
-    } catch (e) {
-      vscode.window.showErrorMessage("Nexus IQ extension: " + e);
-      return;
-    }
-  }
-  private async getApplicationId(applicationPublicId: string) {
-    console.log("getApplicationId", applicationPublicId);
-    return new Promise((resolve, reject) => {
-      request.get(
-        {
-          method: "GET",
-          url: `${
-            this.url
-          }/api/v2/applications?publicId=${applicationPublicId}`,
-          auth: { user: this.user, pass: this.password }
-        },
-        (err:any, response:any, body:any) => {
-          if (err) {
-            reject(`Unable to retrieve Application ID: ${err}`);
-            return;
-          }
-          resolve(body);
-          return;
-        }
-      );
-    });
-  }
-
-  private async submitToIqForEvaluation(data: any, applicationInternalId: string) {
-    return new Promise((resolve, reject) => {
-      request.post(
-        {
-          method: "POST",
-          url: `${
-            this.url
-          }/api/v2/evaluation/applications/${applicationInternalId}`,
-          json: data,
-          auth: { user: this.user, pass: this.password }
-        },
-        (err:any, response:any, body:any) => {
-          if (err) {
-            reject(`Unable to perform IQ scan: ${err}`);
-            return;
-          }
-          let resultId = body.resultId;
-          resolve(resultId);
-          return;
-        }
-      );
-    });
-  }
-
-  private async asyncPollForEvaluationResults(
-    applicationInternalId: string,
-    resultId: string
-  ) {
-    return new Promise((resolve, reject) => {
-      this.pollForEvaluationResults(
-        applicationInternalId,
-        resultId,
-        body => resolve(body),
-        (statusCode, message) =>
-          reject(
-            `Could not fetch evaluation result, code ${statusCode}, message ${message}`
-          )
-      );
-    });
-  }
-
-  private pollForEvaluationResults(
-    applicationInternalId: string,
-    resultId: string,
-    success: (body: string) => any,
-    failed: (statusCode: number, message: string) => any
-  ) {
-    let _this = this;
-    let pollAttempts = 0;
-
-    let successHandler = function(value: string) {
-      success(value);
-    };
-    let errorHandler = function(statusCode: number, message: string) {
-      if (statusCode === 404) {
-        // report still being worked on, continue to poll
-        pollAttempts += 1;
-        // TODO use the top-level class constant, but references are failing
-        if (pollAttempts >= _this.getmaximumEvaluationPollAttempts) {
-          failed(statusCode, "Poll limit exceeded, try again later");
-        } else {
-          setTimeout(() => {
-            _this.getEvaluationResults(
-              applicationInternalId,
-              resultId,
-              successHandler,
-              errorHandler
-            );
-          }, _this.evaluationPollDelayMs);
-        }
-      } else {
-        failed(statusCode, message);
-      }
-    };
-    this.getEvaluationResults(
-      applicationInternalId,
-      resultId,
-      successHandler,
-      errorHandler
-    );
-  }
-
-  private getEvaluationResults(
-    applicationInternalId: string,
-    resultId: string,
-    resolve: (body: string) => any,
-    reject: (statusCode: number, message: string) => any
-  ) {
-    request.get(
-      {
-        method: "GET",
-        url: `${
-          this.url
-        }/api/v2/evaluation/applications/${applicationInternalId}/results/${resultId}`,
-        auth: { user: this.user, pass: this.password }
-      },
-      (error:any, response:any, body:any) => {
-        if (response && response.statusCode != 200) {
-          reject(response.statusCode, error);
-          return;
-        }
-        if (error) {
-          reject(response.statusCode, error);
-          return;
-        }
-        resolve(body);
-      }
-    );
-  }
-
-  private pathExists(p: string): boolean {
-    try {
-      fs.accessSync(p);
-      return true;
-    } catch (err) {
-      return false;
-    }
-  }
-}
+import { IqComponentModel } from "./IqComponentModel";
 
 export class NexusExplorerProvider
   implements vscode.TreeDataProvider<ComponentEntry> {
@@ -331,27 +39,10 @@ export class NexusExplorerProvider
     readonly componentModel: IqComponentModel
   ) {
     this.refresh();
-    // this.reloadComponentModel();
-    // this.autoRefresh = vscode.workspace.getConfiguration('nexusExplorer').get('autorefresh');
-    // vscode.workspace.onDidChangeConfiguration(() => {
-    // 	this.autoRefresh = vscode.workspace.getConfiguration('nexusExplorer').get('autorefresh');
-    // });
   }
-
-  // sortByVulnerability(offset?: number): void {
-  // 	this.reloadComponentModel().then(v => {
-  // 		if (this.componentModel.components.length > 0) {
-  // 			console.log('Sort By Vulnerability');
-  // 			//make a copy of the nodes and then sort the items
-  // 			//using the policy provider
-  // 		}
-  // 	});
-  // }
-
   refresh(offset?: number): void {
     this.reloadComponentModel().then(v => {
       if (this.componentModel.components.length > 0) {
-        //this._onDidChangeTreeData.fire(this.getTreeItem(this.getChildren()[0]));
         this._onDidChangeTreeData.fire();
       }
     });
@@ -365,9 +56,6 @@ export class NexusExplorerProvider
     var tree = dependencyTree({
       filename: path.join("", entry.name),
       directory: ".",
-      //   requireConfig: "path/to/requirejs/config", // optional
-      //   webpackConfig: "path/to/webpack/config", // optional
-      //   tsConfig: "path/to/typescript/config", // optional
       nodeModulesConfig: {
         entry: "module"
       }, // optional
@@ -392,14 +80,12 @@ export class NexusExplorerProvider
       arguments: [entry]
     };
     let maxThreat = entry.maxPolicy();
-    // let tree = this.getDependencies(entry);
     // TODO flesh out more details in the tooltip?
     treeItem.tooltip = `Name: ${entry.name}
-Version: ${entry.version}
-Hash: ${entry.hash}
-Policy: ${maxThreat}
+      Version: ${entry.version}
+      Hash: ${entry.hash}
+      Policy: ${maxThreat}`;
 
-`;
     return treeItem;
   }
 
@@ -420,7 +106,6 @@ Policy: ${maxThreat}
 export class NexusExplorer {
   private nexusViewer: vscode.TreeView<ComponentEntry>;
   private componentModel: IqComponentModel;
-  //private componentModel : IqStaticComponentModel;
   private nexusExplorerProvider: NexusExplorerProvider;
 
   constructor(readonly context: vscode.ExtensionContext) {
@@ -433,10 +118,6 @@ export class NexusExplorer {
     let maximumEvaluationPollAttempts = parseInt(
       config.get("maximumEvaluationPollAttempts") + "", 10);
 
-    /////////////////////////
-    //let applicationId = config.get("applicationId") + '';//.toString();
-    // let applicationId = "XXXXXX";//await this.getApplicationId(applicationPublicId);
-
     /* Please note that login information is hardcoded only for this example purpose and recommended not to do it in general. */
     this.componentModel = new IqComponentModel(
       url,
@@ -445,7 +126,7 @@ export class NexusExplorer {
       applicationPublicId,
       maximumEvaluationPollAttempts
     );
-    //this.componentModel = new IqStaticComponentModel();
+
     this.nexusExplorerProvider = new NexusExplorerProvider(
       context,
       this.componentModel
@@ -455,13 +136,9 @@ export class NexusExplorer {
       treeDataProvider: this.nexusExplorerProvider
     });
 
-    //this.reveal();
-
-    //vscode.window.registerTreeDataProvider('nexusExplorer', this.nexusExplorerProvider);
     vscode.commands.registerCommand("nexusExplorer.refresh", () =>
       this.nexusExplorerProvider.refresh()
     );
-    // vscode.commands.registerCommand('nexusExplorer.sortByVulnerability', () => this.nexusExplorerProvider.sortByVulnerability());
 
     vscode.commands.registerCommand("nexusExplorer.revealResource", () =>
       this.reveal()
